[social] refactoring toot - tooting only
